# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an api for the language learning app",
        "title": "Language Learning App",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/resources/exercise/dialogue/continue": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Continue with dialogue Exercise based on the input lesson",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercise"
                ],
                "summary": "Continue with Dialogue Exercise",
                "parameters": [
                    {
                        "description": "Exercise Request object to be generated",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GenerateDialogueContinuationExerciseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agents.GeneratedDialogueContinuationExercise"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/resources/exercise/dialogue/init": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Init dialogue Exercise based on the input lesson",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercise"
                ],
                "summary": "Init Dialogue Exercise",
                "parameters": [
                    {
                        "description": "Exercise Request object to be generated",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GenerateDialogueInitExerciseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agents.GeneratedDialogueInitExercise"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/resources/exercise/reading-comprehension/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a new reading comprehension Exercise based on the input lesson",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercise"
                ],
                "summary": "Get Reading Comprehension Exercise",
                "parameters": [
                    {
                        "description": "Exercise Request object to be generated",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GenerateReadingComprehensionExerciseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agents.GeneratedReadingComprehensionExercise"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/resources/exercise/reading-comprehension/grade": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Grade a Reading Comprehension Exercise based on the input lesson and user's response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercise"
                ],
                "summary": "Grade Reading Comprehension Exercise",
                "parameters": [
                    {
                        "description": "Exercise Request object to be graded",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GradeReadingComprehensionResponseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/agents.UsageGrade"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/resources/exercise/translation/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a new Translation Exercise based on the input lesson",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercise"
                ],
                "summary": "Get Translation Exercise",
                "parameters": [
                    {
                        "description": "Exercise Request object to be generated",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GenerateTranslationExerciseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agents.GeneratedTranslationExercise"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/resources/exercise/translation/grade": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Grade a Translation Exercise based on the input lesson",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercise"
                ],
                "summary": "Grade Translation Exercise",
                "parameters": [
                    {
                        "description": "Exercise Request object to be graded",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GradeTranslationExerciseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/agents.UsageGrade"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/resources/exercise/usage/grade": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Grade the usage of the language in a sentence",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exercise"
                ],
                "summary": "Grade Usage of language",
                "parameters": [
                    {
                        "description": "Grade Usage Request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GradeUsageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/agents.UsageGrade"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "agents.GeneratedDialogueContinuationExercise": {
            "type": "object",
            "properties": {
                "next": {
                    "type": "string"
                }
            }
        },
        "agents.GeneratedDialogueInitExercise": {
            "type": "object",
            "properties": {
                "init": {
                    "type": "string"
                },
                "situation": {
                    "type": "string"
                }
            }
        },
        "agents.GeneratedReadingComprehensionExercise": {
            "type": "object",
            "properties": {
                "questions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "short_text": {
                    "type": "string"
                }
            }
        },
        "agents.GeneratedTranslationExercise": {
            "type": "object",
            "properties": {
                "sentences": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "agents.UsageGrade": {
            "type": "object",
            "properties": {
                "feedback": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
...
        "dto.GenerateDialogueContinuationExerciseRequest": {
            "type": "object",
            "properties": {
                "history": {
                    "type": "string"
                },
                "lessonId": {
                    "type": "integer"
                }
            }
        },
        "dto.GenerateDialogueInitExerciseRequest": {
            "type": "object",
            "properties": {
                "lessonId": {
                    "type": "integer"
                }
            }
        },
...
        "dto.GenerateReadingComprehensionExerciseRequest": {
            "type": "object",
            "properties": {
                "lessonId": {
                    "type": "integer"
                }
            }
        },
        "dto.GenerateTranslationExerciseRequest": {
            "type": "object",
            "properties": {
                "lessonId": {
                    "type": "integer"
                }
            }
        },
...
        "dto.GradeReadingComprehensionResponseRequest": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "lessonId": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "shortText": {
                    "type": "string"
                }
            }
        },
        "dto.GradeTranslationExerciseRequest": {
            "type": "object",
            "properties": {
                "lessonId": {
                    "type": "integer"
                },
                "originalSentence": {
                    "type": "string"
                },
                "translatedSentence": {
                    "type": "string"
                }
            }
        },
        "dto.GradeUsageRequest": {
            "type": "object",
            "properties": {
                "grammarPatternOrWord": {
                    "type": "string"
                },
                "sentece": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer \u003ctoken\u003e' to correctly set the JWT",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}